shorting, nesting, and using 2013 semantic in the flexbox.less
I tried to simplify with nested mixins. Shorten the code and easy to modify.
The mixins use the 2013 syntax and it degrades it self (when possible) to 2009 an tweener syntax. 

I have split in container/item mixins.

I added two main mixins that can add all the rules at once, and adds nothing if you ommit a value